[ADD] stock_picking_responsible: split from beesdoo_inventory
