+++ beesdoo_inventory/__manifest__.py
@@ -13,7 +13,12 @@
     "website": "https://github.com/beescoop/Obeesdoo",
     "category": "Inventory",
     "version": "12.0.1.0.1",
-    "depends": ["delivery", "beesdoo_base", "beesdoo_product"],
+    "depends": [
+        "delivery",
+        "beesdoo_base",
+        "beesdoo_product",
+        "stock_picking_responsible",
+    ],
     "data": ["views/stock.xml"],
     "installable": True,
     "license": "AGPL-3",
