+++ beesdoo_inventory/readme/DESCRIPTION.rst
@@ -1,1 +1,1 @@
-Adds a responsible, a max shipping date and a button to copy quantity to stock pickings.
+Adds a max shipping date and a button to copy quantity to stock pickings.
