+++ beesdoo_inventory/views/stock.xml
@@ -1,15 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record model="ir.ui.view" id="beesdoo_stock_tree_view">
-        <field name="name">beesdoo.stock.tree.view</field>
-        <field name="model">stock.picking</field>
-        <field name="inherit_id" ref="stock.vpicktree" />
-        <field name="arch" type="xml">
-            <field name="partner_id" position="after">
-                <field name="responsible" />
-            </field>
-        </field>
-    </record>
 
     <record model="ir.ui.view" id="beesdoo_stock_form_view">
         <field name="name">beesdoo.stock.form.view</field>
@@ -18,9 +8,6 @@
         <field name="arch" type="xml">
             <field name="scheduled_date" position="after">
                 <field name="max_shipping_date" placeholder="Max shipping date" />
-            </field>
-            <field name="backorder_id" position="after">
-                <field name="responsible" />
             </field>
             <button name="action_assign" position="after">
                 <button
