+++ setup/stock_picking_responsible/odoo/addons/stock_picking_responsible
@@ -1,0 +1,1 @@
+../../../../stock_picking_responsible
