+++ stock_picking_responsible/README.rst
@@ -1,0 +1,69 @@
+=========================
+Stock Picking Responsible
+=========================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-beescoop%2Fobeesdoo-lightgray.png?logo=github
+    :target: https://github.com/beescoop/obeesdoo/tree/12.0/stock_picking_responsible
+    :alt: beescoop/obeesdoo
+
+|badge1| |badge2| |badge3| 
+
+Adds a responsible partner to the picking and subscribe them
+to notifications.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/beescoop/obeesdoo/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/beescoop/obeesdoo/issues/new?body=module:%20stock_picking_responsible%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* BEES coop SC
+* Coop IT Easy SC
+
+Contributors
+~~~~~~~~~~~~
+
+* `BEES coop SC <http://bees-coop.be>`_ :
+
+  * Augustin Borsu
+  * Elise Dupont
+  * Thibault François
+  * Jean-Marc François
+
+* `Coop IT Easy SCRLfs <https://coopiteasy.be>`_:
+
+  * Houssine Bakkali
+  * Robin Keunen
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `beescoop/obeesdoo <https://github.com/beescoop/obeesdoo/tree/12.0/stock_picking_responsible>`_ project on GitHub.
+
+You are welcome to contribute.
