+++ stock_picking_responsible/__manifest__.py
@@ -1,0 +1,16 @@
+# Copyright 2017 - 2021 BEES coop SCRLfs
+# Copyright 2022 - Coop IT Easy SC
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+{
+    "name": "Stock Picking Responsible",
+    "summary": "Adds a responsible partner to the picking"
+    " and subscribe them to notifications.",
+    "author": "BEES coop SC, Coop IT Easy SC",
+    "website": "https://github.com/beescoop/Obeesdoo",
+    "category": "Inventory",
+    "version": "12.0.1.0.1",
+    "depends": ["stock"],
+    "data": ["views/stock_picking.xml"],
+    "installable": True,
+    "license": "AGPL-3",
+}
