+++ stock_picking_responsible/models/__init__.py
@@ -1,0 +1,1 @@
+from . import stock_picking
