+++ stock_picking_responsible/models/stock_picking.py
@@ -1,0 +1,53 @@
+# Copyright 2022 Coop IT Easy SC
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+
+from odoo import api, fields, models
+
+
+class StockPicking(models.Model):
+    _inherit = "stock.picking"
+
+    responsible = fields.Many2one(
+        comodel_name="res.partner",
+        string="Responsible",
+        default=lambda self: self.env.user.partner_id.id,
+    )
+
+    @api.multi
+    def _add_follower(self):
+        for picking in self:
+            if picking.responsible:
+                types = self.env["mail.message.subtype"].search(
+                    [
+                        "|",
+                        ("res_model", "=", "stock.picking"),
+                        ("name", "=", "Discussions"),
+                    ]
+                )
+                if not self.env["mail.followers"].search(
+                    [
+                        ("res_id", "=", picking.id),
+                        ("res_model", "=", "stock.picking"),
+                        ("partner_id", "=", picking.responsible.id),
+                    ]
+                ):
+                    self.env["mail.followers"].create(
+                        {
+                            "res_model": "stock.picking",
+                            "res_id": picking.id,
+                            "partner_id": picking.responsible.id,
+                            "subtype_ids": [(6, 0, types.ids)],
+                        }
+                    )
+
+    @api.multi
+    def write(self, values):
+        res = super(StockPicking, self).write(values)
+        self._add_follower()
+        return res
+
+    @api.model
+    def create(self, values):
+        picking = super(StockPicking, self).create(values)
+        picking._add_follower()
+        return picking
