+++ stock_picking_responsible/readme/CONTRIBUTORS.rst
@@ -1,0 +1,11 @@
+* `BEES coop SC <http://bees-coop.be>`_ :
+
+  * Augustin Borsu
+  * Elise Dupont
+  * Thibault François
+  * Jean-Marc François
+
+* `Coop IT Easy SCRLfs <https://coopiteasy.be>`_:
+
+  * Houssine Bakkali
+  * Robin Keunen
