+++ stock_picking_responsible/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+Adds a responsible partner to the picking and subscribe them
+to notifications.
