+++ stock_picking_responsible/views/stock_picking.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="beesdoo_stock_tree_view">
+        <field name="name">beesdoo.stock.tree.view</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.vpicktree" />
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="responsible" />
+            </field>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="beesdoo_stock_form_view">
+        <field name="name">beesdoo.stock.form.view</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_form" />
+        <field name="arch" type="xml">
+            <field name="backorder_id" position="after">
+                <field name="responsible" />
+            </field>
+        </field>
+    </record>
+</odoo>
